requirements-ultra-basic.txt
Flask==2.3.2
flask-cors==4.0.0
Pillow==9.5.0
pytesseract==0.3.10
openpyxl==3.1.2
